Moved shared and kernel libs. Added cmake lists for kernel libs
