+++ src/kernel/boot/tss/task_state_segment.cpp
@@ -16,7 +16,7 @@
  */
 
 #include <cstdint>
-#include <kernel/lib/libc/string.h>
+#include <string.h>
 
 #include "task_state_segment.h"
 
@@ -35,7 +35,7 @@ namespace kernel::boot::tss {
      */
     void initializeTaskStateSegment(uint32_t stackPointer) {
         // Clear TSS memory to all zeroes.
-        lib::memset(&tssEntry, 0, sizeof(tssEntry));
+        memset(&tssEntry, 0, sizeof(tssEntry));
 
         //
         tssEntry = {
