--- src/kernel/lib/CMakeLists.txt
@@ -1,3 +1,0 @@
-
-#add_library(kernel_libraries libcpp/pure_function.cpp)
-#target_include_directories(kernel_libraries PUBLIC .)
