+++ src/kernel/library/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(LibC)
+add_subdirectory(LibCpp)
+add_subdirectory(LibDebug)
