+++ src/kernel/library/LibC/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(KLibC INTERFACE)
+target_link_libraries(KLibC INTERFACE sharedLibC)
+target_include_directories(KLibC INTERFACE include)
