+++ src/kernel/library/LibCpp/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(KLibCpp lib/pure_function.cpp)
+target_include_directories(KLibCpp PRIVATE include)
