+++ src/kernel/library/LibCpp/lib/pure_function.cpp
