+++ src/kernel/library/LibDebug/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(KLibDebug lib/debug.asm)
+target_include_directories(KLibDebug PRIVATE include)
