+++ src/shared/library/LibC/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(sharedLibC lib/string.cpp)
+target_include_directories(sharedLibC PRIVATE include)
